Add ort pipelines, their docs and diffusion pipelines docs (#41)
I tried to fix the image processor loading issue I encountered in #34
(image-to-text task) but It was difficult without copying a big chunk of
the preprocessors loading logic in transformers.pipelines.pipeline.
In this approach we only need to handle the task to class mapping which
is simpler than handling the preprocessors loading.
I also added some specific docs for pipelines, mostly linking to
transformers.pipeline docs and fixed diffusion docs not showing __call__
signature/params/docstring (check
[here](https://moon-ci-docs.huggingface.co/docs/optimum-onnx/pr_41/en/onnxruntime/package_reference/modeling_diffusion#optimum.onnxruntime.ORTStableDiffusionPipeline.__call__))